move components to separate folders
