+++ src/class-form/ClassForm.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import {TextField, Checkbox} from 'office-ui-fabric-react'
-import SchemaForm from "./SchemaForm"
+import SchemaForm from "../schema-form/SchemaForm"
 import faunadb from 'faunadb';
-import clientForSubDB from "./clientForSubDB";
+import clientForSubDB from "../clientForSubDB";
 const q = faunadb.query, Ref = q.Ref;
 
 export class ClassForm extends Component {
